+++ packages/core/src/index.ts
@@ -640,15 +640,15 @@ class Hopr extends EventEmitter {
   // @TODO make modules Startable
   public async stop(): Promise<void> {
     if (this.status == 'DESTROYED') {
-      throw Error(`alreayd destroyed. Cannot destroy twice`)
+      throw Error(`Hopr instance already destroyed.`)
     }
     this.status = 'DESTROYED'
     verbose('Stopping checking timeout')
     this.stopPeriodicCheck?.()
     verbose('Stopping heartbeat & indexer')
-    await this.heartbeat.stop()
+    this.heartbeat?.stop()
     verbose(`Stopping connector`)
-    await this.connector.stop()
+    await this.connector?.stop()
     verbose('Stopping database')
     await this.db?.close()
     log(`Database closed.`)
@@ -1195,6 +1195,10 @@ class Hopr extends EventEmitter {
 
   public async getPublicKeyOf(addr: Address): Promise<PublicKey> {
     return await this.connector.getPublicKeyOf(addr)
+  }
+
+  public async getEntryNodes(): Promise<{ id: PeerId; multiaddrs: Multiaddr[] }[]> {
+    return this.connector.waitForPublicNodes()
   }
 
   // @TODO remove this