+++ packages/hoprd/hopr-admin/src/commands/withdraw.ts
@@ -1,19 +1,12 @@
 import type API from '../utils/api'
 import { utils as ethersUtils } from 'ethers'
-import { Command, type CacheFunctions } from '../utils/command'
+import { Command, type CacheFunctions, type HoprOrNative } from '../utils/command'
 
 export default class Withdraw extends Command {
   constructor(api: API, cache: CacheFunctions) {
     super(
       {
-        default: [
-          [
-            ['number', 'amount to withdraw', false],
-            ['hoprOrNative', 'withdraw "HOPR" or "NATIVE"', false],
-            ['nativeAddress', 'recipient', false]
-          ],
-          'withdraw'
-        ]
+        default: [[['number', 'amount'], ['hoprOrNative'], ['nativeAddress']], 'withdraw']
       },
       api,
       cache
@@ -36,14 +29,22 @@ export default class Withdraw extends Command {
       string | undefined,
       string,
       number,
-      string,
+      HoprOrNative,
       string
     ]
     if (error) return log(error)
 
     const amountWei = ethersUtils.parseEther(String(amount))
     const response = await this.api.withdraw(amountWei.toString(), currency, recipient)
-    if (!response.ok) return log(this.invalidResponse('withdraw'))
+
+    if (!response.ok) {
+      return log(
+        await this.failedApiCall(response, 'withdraw', {
+          400: (v) => `one or more invalid inputs ${v.status}`,
+          422: (v) => v.status
+        })
+      )
+    }
 
     const receipt = response.json().then((res) => res.receipt)
     return log(`Withdrawing ${amount} ${currency} to ${recipient}, receipt ${receipt}.`)