+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Warn when manual path selection contains duplicate adjacent entries
 - Correctly recognize Avado/Dappnode private subnet ([#4032](https://github.com/hoprnet/hoprnet/pull/4032))
 - Redesign entry node code to recycle knowledge and automatically reconnect ([#3990]](https://github.com/hoprnet/hoprnet/pull/3990))
+- Add `entryNodes` command to API and `hopr-admin` ([#4049](https://github.com/hoprnet/hoprnet/pull/4049))
 
 # Breaking changes